pyproject.toml
[tool.poetry]
name = "budget"
version = "0.1.0"
description = "simple budget app for personal use"
authors = ["VF"]
readme = "README.md"
packages = [
    { include = "*" },
]

[tool.poetry.dependencies]
python = "^3.11"
django = '4.2.4'
django-split-settings = "^1.2.0"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"
drf-spectacular = "^0.27.1"
django-extra-checks = "^0.14.0"
psycopg2 = "^2.9.9"
pandas = "^2.2.1"
chardet = "^5.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "1.7.1"
ruff = "^0.3.3"
pre-commit = "^3.6.2"
django-stubs = "^4.2.7"
pandas-stubs = "^2.2.1.240316"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict_optional = 'True'
python_version = "3.11"
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
follow_imports = "normal"
ignore_missing_imports = true


[mypy.plugins.django-stubs]
django_settings_module = "backend.server.settings"
strict_settings = false
ignore_missing_settings = true
ignore_missing_model_attributes = true

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "backend.server.settings"
